+++ src/ByteDev.Configuration.Environment/EnvironmentVariableProvider.cs
@@ -572,7 +572,7 @@ namespace ByteDev.Configuration.Environment
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="format" /> is null.</exception>
         /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
-        /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a defined name or number value of the enum.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a DateTime based on the supplied format.</exception>
         public DateTime GetDateTime(string name, string format)
         {
             if (format == null)
@@ -587,6 +587,33 @@ namespace ByteDev.Configuration.Environment
             catch (FormatException ex)
             {
                 throw new UnexpectedEnvironmentVariableTypeException(name, value, typeof(DateTime), ex);
+            }
+        }
+
+        /// <summary>
+        /// Retrieve an environment variable as a DateTime. If it does not exist or it's value cannot be cast 
+        /// based on the supplied format then the <paramref name="defaultValue" /> will be returned.
+        /// </summary>
+        /// <param name="name">Name of the environment variable.</param>
+        /// <param name="format">Format of the string.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        /// <exception cref="T:System.ArgumentNullException"><paramref name="format" /> is null.</exception>
+        public DateTime GetDateTimeOrDefault(string name, string format, DateTime defaultValue)
+        {
+            if (format == null)
+                throw new ArgumentNullException(nameof(format));
+
+            var value = GetStringOrDefault(name);
+
+            try
+            {
+                return DateTime.ParseExact(value, format, null, DateTimeStyles.None);
+            }
+            catch (Exception)
+            {
+                return defaultValue;
             }
         }
     }